test/webtests/http/tests/devtools/a11y-axe-core/application-panel/clear-storage-a11y-test-expected.txt

CONSOLE MESSAGE: InspectorTest.IndexedDB_callback1
CONSOLE MESSAGE: InspectorTest.IndexedDB_callback2
CONSOLE MESSAGE: InspectorTest.IndexedDB_callback3
Tests accessibility in the Clear Storage view using the axe-core linter.
Clear storage view is visible: true
aXe violations: []
